backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
python-dotenv==1.0.0
requests==2.31.0
pillow==10.0.0
